Remove some magic_enum from headers
