+++ src/Common/IntervalKind.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <base/types.h>
-#include <base/EnumReflection.h>
 
 namespace DB
 {
@@ -27,7 +26,7 @@ struct IntervalKind
     IntervalKind(Kind kind_ = Kind::Second) : kind(kind_) {} /// NOLINT
     operator Kind() const { return kind; } /// NOLINT
 
-    constexpr std::string_view toString() const { return magic_enum::enum_name(kind); }
+    std::string_view toString() const;
 
     /// Returns number of nanoseconds in one interval.
     /// For `Month`, `Quarter` and `Year` the function returns an average number of nanoseconds.
